Replace Producer/Consumer with IData-types
